[hyeok][refactor] Plan Side Bar detail button and search button refactoring
